busca_rota_django/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
psycopg2-binary
drf-yasg

